requirements.txt
astroid==1.5.3
backports.functools-lru-cache==1.4
certifi==2017.4.17
chardet==3.0.4
configparser==3.5.0
Django==1.8
dnspython==1.15.0
email-validator==1.0.2
enum34==1.1.6
factory-boy==2.9.2
Faker==0.8.3
idna==2.5
ipaddress==1.0.18
isort==4.2.15
lazy-object-proxy==1.3.1
mccabe==0.6.1
page-objects==1.1.0
pep8==1.7.0
pylint==1.7.2
python-dateutil==2.6.1
requests==2.18.1
selenium==2.53.6
singledispatch==3.4.0.3
six==1.10.0
urllib3==1.21.1
wrapt==1.10.10
